+++ app/models/research_session/steps.rb
@@ -11,7 +11,7 @@ class ResearchSession
       topic:         [:topic, :purpose],
       methodologies: [:other_methodology, methodologies: []],
       recording:     [:other_recording_method, recording_methods: []],
-      storing:       [:shared_with, :shared_duration, :shared_use],
+      storing:       [:shared_with, :shared_duration],
       where_when:    [:where_when_enabled, :when_text, :duration, :location, :participant_equipment,
                       :food_provided],
       expenses:      [:expenses_enabled, :travel_expenses_limit, :food_expenses_limit,