+++ db/migrate/20171116154338_remove_shared_use.rb
@@ -1,0 +1,5 @@
+class RemoveSharedUse < ActiveRecord::Migration[5.1]
+  def change
+    remove_column :research_sessions, :shared_use, :string
+  end
+end