+++ spec/factories/research_session.rb
@@ -35,7 +35,6 @@ FactoryBot.define do
       status :storing
       shared_with :team
       shared_duration '1 year'
-      shared_use 'To train others'
     end
 
     trait :step_where_when do