+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [FEATURE] - [Researcher page: we should have job title of the lead researcher as a field](https://trello.com/c/HvKi9NwT/169-2-researcher-page-we-should-have-job-title-of-the-lead-researcher-as-a-field)
 * [BUG] - [When a sterling sign is added in, the preview generates 0.00](https://trello.com/c/IHdoFxEl/246-3-bug-when-a-sterling-sign-is-added-in-the-preview-generates-000)
 * [FEATURE] - [Create a New Form and Info Sheet from a Saved Form and Info Sheet (aka Research Session)](https://trello.com/c/qYBbtDC9/87-create-a-new-form-and-info-sheet-from-a-saved-form-and-info-sheet-aka-research-session)
+* [FEATURE] - [GDPR compliance for our Data (now Storing) step](https://trello.com/c/kkuQxS8T/224-gdpr-compliance-for-our-data-now-storing-step)
 
 # 0.4.0 / 2017-11-07